+++ src/main/java/com/jzy/service/impl/MissLessonStudentServiceImpl.java
@@ -8,9 +8,12 @@ import com.jzy.model.dto.MissLessonStudentDetailedDto;
 import com.jzy.model.dto.MissLessonStudentSearchCondition;
 import com.jzy.model.dto.MyPage;
 import com.jzy.service.MissLessonStudentService;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -22,6 +25,8 @@ import java.util.List;
  **/
 @Service
 public class MissLessonStudentServiceImpl extends AbstractServiceImpl implements MissLessonStudentService {
+    private final static Logger logger = LogManager.getLogger(MissLessonStudentServiceImpl.class);
+
     @Autowired
     private MissLessonStudentMapper missLessonStudentMapper;
 
@@ -34,7 +39,7 @@ public class MissLessonStudentServiceImpl extends AbstractServiceImpl implements
 
     @Override
     public String updateMissLessonStudentInfo(MissLessonStudentDetailedDto missLessonStudentDetailedDto) {
-        if (missLessonStudentDetailedDto.getCurrentClassId().equals(missLessonStudentDetailedDto.getOriginalClassId())){
+        if (missLessonStudentDetailedDto.getCurrentClassId().equals(missLessonStudentDetailedDto.getOriginalClassId())) {
             //原班级和补课班级相同
             return "theSameClass";
         }
@@ -49,13 +54,13 @@ public class MissLessonStudentServiceImpl extends AbstractServiceImpl implements
             return "currentClassNotExist";
         }
 
-        missLessonStudentMapper.missLessonStudentDetailedDto(missLessonStudentDetailedDto);
+        missLessonStudentMapper.updateMissLessonStudentInfo(missLessonStudentDetailedDto);
         return Constants.SUCCESS;
     }
 
     @Override
     public String insertMissLessonStudent(MissLessonStudentDetailedDto missLessonStudentDetailedDto) {
-        if (missLessonStudentDetailedDto.getCurrentClassId().equals(missLessonStudentDetailedDto.getOriginalClassId())){
+        if (missLessonStudentDetailedDto.getCurrentClassId().equals(missLessonStudentDetailedDto.getOriginalClassId())) {
             //原班级和补课班级相同
             return "theSameClass";
         }
@@ -75,12 +80,30 @@ public class MissLessonStudentServiceImpl extends AbstractServiceImpl implements
     }
 
     @Override
-    public void deleteOneMissLessonStudentById(Long id) {
-        missLessonStudentMapper.deleteOneMissLessonStudentById(id);
+    public long deleteOneMissLessonStudentById(Long id) {
+        if (id == null) {
+            return 0;
+        }
+        return missLessonStudentMapper.deleteOneMissLessonStudentById(id);
     }
 
     @Override
-    public void deleteManyMissLessonStudentsByIds(List<Long> ids) {
-        missLessonStudentMapper.deleteManyMissLessonStudentsByIds(ids);
+    public long deleteManyMissLessonStudentsByIds(List<Long> ids) {
+        if (ids == null ||ids.size() == 0){
+            return 0;
+        }
+        return missLessonStudentMapper.deleteManyMissLessonStudentsByIds(ids);
+    }
+
+    @Override
+    public String deleteMissLessonStudentsByCondition(MissLessonStudentSearchCondition condition) {
+        List<MissLessonStudentDetailedDto> dtos=missLessonStudentMapper.listMissLessonStudents(condition);
+        List<Long> ids=new ArrayList<>();
+        for (MissLessonStudentDetailedDto dto:dtos){
+            ids.add(dto.getId());
+        }
+
+        deleteManyMissLessonStudentsByIds(ids);
+        return Constants.SUCCESS;
     }
 }
